pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "offerdoc-generator"
version = "1.0.0"
description = "Document generation tool"
readme = "README.md"
requires-python = ">=3.13"
license = "MIT"
keywords = ["document", "generator", "docx", "template"]
authors = [
  { name = "Your Name", email = "your.email@example.com" },
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "python-docx>=1.1.2",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "jinja2>=3.1.5",
    "PyYAML>=6.0.2",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "black>=25.1.0",
    "isort>=6.0.0",
    "mypy>=1.15.0",
    "ruff>=0.9.6",
]

[tool.hatch.version]
path = "src/odg/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/odg"]

[project.scripts]
odg = "odg.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--strict-markers"
markers = [
    "end_to_end: mark tests as end-to-end integration tests",
]

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]
extend-select = ["C90", "UP", "ANN", "ICN", "PIE"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_any_unimported = true
